requirements.txt
Flask
deepl
python-dotenv
werkzeug
yandexcloud
